Moved some files in `src/lib/components/explorer`
closes #263

Also:
- deleted `fe/src/lib/components/svizzle/utils/shared.js` which only had `export const defaultRel = 'noopener';`,  unused elsewhere
