+++ fe/src/lib/components/explorer/banners/OrgBanner.svelte
@@ -1,8 +1,7 @@
 <script>
 	import {_screen, Banner, LayoutHMF} from '@svizzle/ui';
 
-	import OrgDetails from '$lib/components/orgs/OrgDetails.svelte';
-
+	import OrgDetails from '$lib/components/explorer/orgs/OrgDetails.svelte';
 	import {bannersDefaultFooterText} from '$lib/config';
 	import {_hero, _isCursorOnMap, clearHero} from '$lib/stores/interaction';
 	import {_bannersTheme} from '$lib/stores/theme';
