+++ fe/src/lib/components/explorer/icons/City.svelte
