+++ fe/src/lib/components/explorer/icons/Region.svelte
