+++ fe/src/lib/components/explorer/icons/WikipediaLogo.svelte
