+++ fe/src/lib/components/explorer/map/AutoZoomControl.svelte
