+++ fe/src/lib/components/explorer/map/SvgLayers.svelte
