+++ fe/src/lib/components/explorer/map/SvgMarkers.svelte
