--- fe/src/lib/components/svizzle/utils/shared.js
@@ -1,1 +1,0 @@
-export const defaultRel = 'noopener';
